LayoutTests/http/tests/xmlhttprequest/infoOnProgressEvent-expected.txt

Test case for bug 13596: Implement .onprogress handler on XMLHttpRequest objects to support progressive download content length information

You should see type, bubble, cancelable, eventPhase, target and current target for synchronous and asynchronous cases

Synchronous case:
Type: progress
Bubble: false
Cancelable: true
EventPhase: 2
Target: [object XMLHttpRequest]
Current target: [object XMLHttpRequest]

Asynchronous case:
Type: progress
Bubble: false
Cancelable: true
EventPhase: 2
Target: [object XMLHttpRequest]
Current target: [object XMLHttpRequest]
